requirements.txt
torch
torchvision
numpy
albumentations
Pillow
PyYAML
tqdm
tensorboard
einops
kornia
timm
